+++ build-i18n.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-I18NDUDE=../../bin/i18ndude
+I18NDUDE=./bin/i18ndude
 I18NPATH=plone/app/event
 DOMAIN=plone.app.event
 $I18NDUDE rebuild-pot --pot $I18NPATH/locales/$DOMAIN.pot --create $DOMAIN $I18NPATH