+++ plone/app/event/browser/event_view.py
@@ -1,5 +1,6 @@
 from Products.Five.browser import BrowserView
 from plone.event.interfaces import IEventAccessor
+from plone.event.interfaces import IOccurrence
 
 
 def get_location(accessor):
@@ -25,3 +26,10 @@ class EventView(BrowserView):
         self.context = context
         self.request = request
         self.data = IEventAccessor(context)
+
+    def __call__(self):
+        if IOccurrence.providedBy(self.context):
+            # The transient Occurrence objects cannot be edited. disable the
+            # edit border for them.
+            self.request.set('disable_border', True)
+        return self.index()  # render me.