+++ plone/app/event/tests/test_event_view.py
@@ -24,7 +24,7 @@ class TestEventViewDX(AbstractSampleDataEvents):
         self.assertTrue('2013-05-15' in output)
         self.assertTrue('(Europe/Vienna / UTC200)' in output)
         self.assertTrue('Graz' in output)
-        self.assertTrue('More occurrences of this event' not in output)
+        self.assertTrue('All dates' not in output)
 
     @mock.patch('plone.app.event.base.localized_now', new=PN)
     def test_event_view__recurring(self):
@@ -34,7 +34,7 @@ class TestEventViewDX(AbstractSampleDataEvents):
 
         self.assertTrue('Now Event' in output)
         self.assertTrue('2013-05-05' in output)
-        self.assertTrue('More occurrences of this event' in output)
+        self.assertTrue('All dates' in output)
         self.assertTrue('2013-05-07' in output)
         self.assertTrue('2013-05-09' in output)
         self.assertTrue('http://plone.org' in output)
@@ -47,9 +47,7 @@ class TestEventViewDX(AbstractSampleDataEvents):
         output = view()
 
         self.assertTrue('Now Event' in output)
-        self.assertTrue('2013-05-05' not in output)
-        self.assertTrue('More occurrences of this event' not in output)
+        self.assertTrue('2013-05-05' in output)
         self.assertTrue('2013-05-07' in output)
-        self.assertTrue('2013-05-09' not in output)
-        self.assertTrue('This event is part of a recurring Event.' in output)
+        self.assertTrue('2013-05-09' in output)
         self.assertTrue('http://plone.org' in output)