+++ plone/app/event/tests/test_recurrence.py
@@ -66,6 +66,13 @@ class TestTraversalDX(AbstractSampleDataEvents):
         self.assertTrue(IOccurrence.providedBy(item))
         self.assertEqual(type(self.now_event), type(item.aq_parent))
 
+        # Test attributes of Occurrence
+        self.assertEqual(item.portal_type, 'Occurrence')
+        self.assertEqual(item.id, '2013-05-07')
+        delta = datetime.timedelta(days=2)
+        self.assertEqual(item.start, self.now + delta)
+        self.assertEqual(item.end, self.now + delta + self.duration)
+
     def test_occurrence_accessor(self):
         start = self.now
         end = self.future
@@ -155,17 +162,22 @@ class TestOccurrences(unittest.TestCase):
         view = zope.component.getMultiAdapter(
             (self.portal['interval'], self.request), name='event_summary')
         result = view.next_occurrences
-        # altogether 5 occurrences, but start occurrence is not included
-        self.assertEqual(4, len(result))
+        # altogether 5 occurrences, start occurrence is included
+        self.assertEqual(5, len(result))
 
         view = zope.component.getMultiAdapter(
             (self.portal['many'], self.request), name='event_summary')
-        result = view.next_occurrences
-        self.assertEqual(view.max_occurrences, len(result))
+
+        # Number of shown occurrences should match max_occurrences setting
+        self.assertEqual(len(view.next_occurrences), view.max_occurrences)
+        # num_more_occurrences should return number of remaining occurrences
+        self.assertEqual(
+            view.num_more_occurrences, 1000 - view.max_occurrences)
 """
 
+
 class MockEvent(SimpleItem):
-    """ Mock event"""
+    """Mock event"""
 
 
 class TestRecurrenceSupport(unittest.TestCase):